pyproject.toml
[tool.poetry]
name = "emailbuddy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.15"
langgraph = "^0.2.14"
langchain-community = "^0.2.14"
huggingface-hub = "^0.24.6"
streamlit = "^1.38.0"
secure-smtplib = "0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"